fix: domain과 presentation 모듈을 kotlin lib로 변경
