+++ .idea/misc.xml
@@ -3,11 +3,12 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/zzxcv/AndroidStudioProjects/three-days-android/presentation/home/src/main/res/layout/activity_main.xml" value="0.25729166666666664" />
         <entry key="..\:/Users/zzxcv/AndroidStudioProjects/three-days-android/presentation/src/main/res/layout/activity_main.xml" value="0.1703125" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_7" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
