+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name="com.depromeet.threedays.presentation.MainActivity"
+            android:name="com.depromeet.threedays.home.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
