+++ build.gradle
@@ -15,6 +15,7 @@ plugins {
     id 'com.android.library' version '7.2.1' apply false
     id 'org.jetbrains.kotlin.android' version '1.6.10' apply false
     id "org.sonarqube" version "3.4.0.2513"
+    id 'org.jetbrains.kotlin.jvm' version '1.6.10' apply false
 }
 
 task clean(type: Delete) {
