+++ data/build.gradle
@@ -33,7 +33,6 @@ android {
 
 dependencies {
     implementation(project(":domain"))
-    implementation(project(":presentation"))
 
     implementation(jetpackDeps)
     implementation(coroutines)
