+++ domain/src/main/java/com/depromeet/threedays/domain/MyClass.kt
@@ -1,0 +1,4 @@
+package com.depromeet.threedays.domain
+
+class MyClass {
+}
