+++ presentation/build.gradle
@@ -1,58 +1,9 @@
 plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
-    id "kotlin-kapt"
-    id 'dagger.hilt.android.plugin'
+    id 'java-library'
+    id 'org.jetbrains.kotlin.jvm'
 }
 
-android {
-    hilt {
-        enableAggregatingTask = true
-    }
-}
-
-android {
-    compileSdk 32
-
-    defaultConfig {
-        minSdk 24
-        targetSdk 32
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-    buildFeatures {
-        dataBinding true
-    }
-}
-
-dependencies {
-    //modules
-    implementation(project(":core"))
-    implementation(project(":core-design-system"))
-    implementation(project(":navigator"))
-    implementation(project(":domain"))
-
-    implementation(jetpackDeps)
-    implementation(coroutines)
-
-    implementation deps.hilt.core
-    kapt deps.hilt.compiler
-
-    testImplementation(testDeps)
-    androidTestImplementation(androidTestDeps)
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_7
+    targetCompatibility = JavaVersion.VERSION_1_7
 }
