+++ presentation/home/build.gradle
@@ -1,23 +1,25 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlin-android'
-    id 'kotlin-kapt'
-    id 'kotlin-parcelize'
+    id "kotlin-kapt"
     id 'dagger.hilt.android.plugin'
+}
+
+android {
+    hilt {
+        enableAggregatingTask = true
+    }
 }
 
 android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.depromeet.threedays"
         minSdk 24
         targetSdk 32
-        versionCode 2
-        versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -39,27 +41,19 @@ android {
 }
 
 dependencies {
-    implementation(project(":domain"))
-    implementation(project(":data"))
     implementation(project(":core"))
+    implementation(project(":core-design-system"))
     implementation(project(":navigator"))
-    implementation(project(":presentation:home"))
+    implementation(project(":domain"))
 
     implementation(jetpackDeps)
     implementation(coroutines)
 
-    //implementation platform(deps.firebaseBom)
-    //implementation(firebaseDeps)
-
     implementation deps.hilt.core
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'com.google.android.material:material:1.4.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     kapt deps.hilt.compiler
-
-    implementation deps.glide.core
-    implementation deps.glide.okhttp
-    kapt deps.glide.compiler
-
-    implementation deps.gson
-    implementation deps.splashscreen
 
     testImplementation(testDeps)
     androidTestImplementation(androidTestDeps)
