+++ presentation/home/src/main/java/com/depromeet/threedays/home/MainActivity.kt
@@ -1,0 +1,11 @@
+package com.depromeet.threedays.home
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
