--- presentation/src/main/java/com/depromeet/threedays/presentation/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.depromeet.threedays.presentation
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
