+++ presentation/src/main/java/com/depromeet/threedays/presentation/MyClass.kt
@@ -1,0 +1,4 @@
+package com.depromeet.threedays.presentation
+
+class MyClass {
+}
