+++ settings.gradle
@@ -17,6 +17,7 @@ include ':app'
 include ':core'
 include ':core-design-system'
 include ':data'
-include ':domain'
 include ':navigator'
+include ':domain'
 include ':presentation'
+include ':presentation:home'
